chore(map): move to `map` directory.
