+++ src/lib/map/index.ts
@@ -1,0 +1,2 @@
+export { DataMap } from './data-map.class';
+export { WeakDataMap } from './weak-data-map.class';
